+++ src-opam/dockerfile_distro.ml
@@ -19,12 +19,12 @@
 open Astring
 
 type t = [ 
-  | `Alpine of [ `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9 | `Latest ]
+  | `Alpine of [ `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10 | `Latest ]
   | `CentOS of [ `V6 | `V7 | `Latest ]
-  | `Debian of [ `V9 | `V8 | `V7 | `Stable | `Testing | `Unstable ]
-  | `Fedora of [ `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `Latest ]
+  | `Debian of [ `V10 | `V9 | `V8 | `V7 | `Stable | `Testing | `Unstable ]
+  | `Fedora of [ `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `V30 | `Latest ]
   | `OracleLinux of [ `V7 | `Latest ]
-  | `OpenSUSE of [ `V42_1 | `V42_2 | `V42_3 | `V15_0 | `Latest ]
+  | `OpenSUSE of [ `V42_1 | `V42_2 | `V42_3 | `V15_0 | `V15_1 | `Latest ]
   | `Ubuntu of [ `V12_04 | `V14_04 | `V15_04 | `V15_10 | `V16_04 | `V16_10 | `V17_04 | `V17_10 | `V18_04 | `V18_10 | `V19_04 | `LTS | `Latest ]
 ] [@@deriving sexp]
 
@@ -35,38 +35,40 @@ type status = [
 ] [@@deriving sexp]
 
 let distros = [
-  `Alpine `V3_3; `Alpine `V3_4; `Alpine `V3_5; `Alpine `V3_6; `Alpine `V3_7; `Alpine `V3_8; `Alpine `V3_9; `Alpine `Latest;
+  `Alpine `V3_3; `Alpine `V3_4; `Alpine `V3_5; `Alpine `V3_6; `Alpine `V3_7; `Alpine `V3_8; `Alpine `V3_9; `Alpine `V3_10; `Alpine `Latest;
   `CentOS `V6; `CentOS `V7; `CentOS `Latest;
-  `Debian `V9; `Debian `V8; `Debian `V7;
+  `Debian `V10; `Debian `V9; `Debian `V8; `Debian `V7;
   `Debian `Stable; `Debian `Testing; `Debian `Unstable;
-  `Fedora `V23; `Fedora `V24; `Fedora `V25; `Fedora `V26; `Fedora `V27; `Fedora `V28; `Fedora `V29; `Fedora `Latest;
+  `Fedora `V23; `Fedora `V24; `Fedora `V25; `Fedora `V26; `Fedora `V27; `Fedora `V28; `Fedora `V29; `Fedora `V30; `Fedora `Latest;
   `OracleLinux `V7; `OracleLinux `Latest;
-  `OpenSUSE `V42_1; `OpenSUSE `V42_2; `OpenSUSE `V42_3; `OpenSUSE `V15_0; `OpenSUSE `Latest;
+  `OpenSUSE `V42_1; `OpenSUSE `V42_2; `OpenSUSE `V42_3; `OpenSUSE `V15_0; `OpenSUSE `V15_1; `OpenSUSE `Latest;
   `Ubuntu `V12_04; `Ubuntu `V14_04; `Ubuntu `V15_04; `Ubuntu `V15_10;
   `Ubuntu `V16_04; `Ubuntu `V16_10; `Ubuntu `V17_04; `Ubuntu `V17_10; `Ubuntu `V18_04; `Ubuntu `V18_10; `Ubuntu `V19_04;
   `Ubuntu `Latest; `Ubuntu `LTS ]
   
 let distro_status (d:t) : status = match d with
   | `Alpine ( `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8) -> `Deprecated
-  | `Alpine `V3_9 -> `Active `Tier1
-  | `Alpine `Latest -> `Alias (`Alpine `V3_9)
+  | `Alpine `V3_9 -> `Active `Tier2
+  | `Alpine `V3_10 -> `Active `Tier1
+  | `Alpine `Latest -> `Alias (`Alpine `V3_10)
   | `CentOS `V7 -> `Active `Tier2
   | `CentOS `V6 -> `Deprecated
   | `CentOS `Latest -> `Alias (`CentOS `V7)
   | `Debian `V7 -> `Deprecated
   | `Debian `V8  -> `Active `Tier2
   | `Debian `V9 -> `Active `Tier1
+  | `Debian `V10 -> `Active `Tier1
   | `Debian `Stable -> `Alias (`Debian `V9)
   | `Debian `Testing -> `Active `Tier2
   | `Debian `Unstable -> `Active `Tier2
-  | `Fedora ( `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28) -> `Deprecated
-  | `Fedora `V29 -> `Active `Tier2
-  | `Fedora `Latest -> `Alias (`Fedora `V29)
+  | `Fedora ( `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29) -> `Deprecated
+  | `Fedora `V30 -> `Active `Tier2
+  | `Fedora `Latest -> `Alias (`Fedora `V30)
   | `OracleLinux `V7 -> `Active `Tier2
   | `OracleLinux `Latest -> `Alias (`OracleLinux `V7)
-  | `OpenSUSE `V42_1 | `OpenSUSE `V42_2 -> `Deprecated
-  | `OpenSUSE (`V42_3|`V15_0) -> `Active `Tier2
-  | `OpenSUSE `Latest -> `Alias (`OpenSUSE `V15_0)
+  | `OpenSUSE (`V42_1 | `V42_2 | `V42_3 | `V15_0) -> `Deprecated
+  | `OpenSUSE `V15_1 -> `Active `Tier2
+  | `OpenSUSE `Latest -> `Alias (`OpenSUSE `V15_1)
   | `Ubuntu (`V16_04 | `V18_04 | `V19_04 ) -> `Active `Tier2
   | `Ubuntu ( `V12_04 | `V14_04 | `V15_04 | `V15_10 | `V16_10 | `V17_04 | `V17_10 | `V18_10) -> `Deprecated
   | `Ubuntu `LTS -> `Alias (`Ubuntu `V18_04)
@@ -88,8 +90,8 @@ module OV = Ocaml_version
 
 let distro_arches ov (d:t) =
   match resolve_alias d, ov with
-  | `Debian `V9, ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `X86_64; `Aarch64; `Ppc64le; `Aarch32 ]
-  | `Alpine (`V3_6 | `V3_7 | `V3_8 | `V3_9), ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `X86_64; `Aarch64 ]
+  | `Debian (`V9|`V10), ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `X86_64; `Aarch64; `Ppc64le; `Aarch32 ]
+  | `Alpine (`V3_6 | `V3_7 | `V3_8 | `V3_9 | `V3_10), ov when OV.(compare Releases.v4_05_0 ov) = -1 -> [ `X86_64; `Aarch64 ]
   | `Ubuntu (`V16_04|`V18_04|`V17_10|`V18_10|`V19_04), ov when OV.(compare Releases.v4_05_0 ov) = -1  -> [ `X86_64; `Aarch64 ]
   | _ -> [ `X86_64 ]
 
@@ -115,6 +117,7 @@ let builtin_ocaml_of_distro (d:t) : string option =
   |`Debian `V7 -> Some "3.12.1"
   |`Debian `V8 -> Some "4.01.0"
   |`Debian `V9 -> Some "4.02.3"
+  |`Debian `V10 -> Some "4.05.0"
   |`Ubuntu `V12_04 -> Some "3.12.1"
   |`Ubuntu `V14_04 -> Some "4.01.0"
   |`Ubuntu `V15_04 -> Some "4.01.0"
@@ -133,6 +136,7 @@ let builtin_ocaml_of_distro (d:t) : string option =
   |`Alpine `V3_7 -> Some "4.04.2"
   |`Alpine `V3_8 -> Some "4.06.1"
   |`Alpine `V3_9 -> Some "4.06.1"
+  |`Alpine `V3_10 -> Some "4.07.0"
   |`Fedora `V21 -> Some "4.01.0"
   |`Fedora `V22 -> Some "4.02.0"
   |`Fedora `V23 -> Some "4.02.2"
@@ -142,12 +146,14 @@ let builtin_ocaml_of_distro (d:t) : string option =
   |`Fedora `V27 -> Some "4.05.0"
   |`Fedora `V28 -> Some "4.06.0"
   |`Fedora `V29 -> Some "4.07.0"
+  |`Fedora `V30 -> Some "4.07.0"
   |`CentOS `V6 -> Some "3.11.2"
   |`CentOS `V7 -> Some "4.01.0"
   |`OpenSUSE `V42_1 -> Some "4.02.3"
   |`OpenSUSE `V42_2 -> Some "4.03.0"
   |`OpenSUSE `V42_3 -> Some "4.03.0"
   |`OpenSUSE `V15_0 -> Some "4.05.0"
+  |`OpenSUSE `V15_1 -> Some "4.05.0"
   |`OracleLinux `V7 -> None
   |`Alpine `Latest |`CentOS `Latest |`OracleLinux `Latest
   |`OpenSUSE `Latest |`Ubuntu `LTS | `Ubuntu `Latest
@@ -171,6 +177,7 @@ let tag_of_distro (d:t) = match d with
   |`Debian `Stable -> "debian-stable"
   |`Debian `Unstable -> "debian-unstable"
   |`Debian `Testing -> "debian-testing"
+  |`Debian `V10 -> "debian-10"
   |`Debian `V9 -> "debian-9"
   |`Debian `V8 -> "debian-8"
   |`Debian `V7 -> "debian-7"
@@ -187,6 +194,7 @@ let tag_of_distro (d:t) = match d with
   |`Fedora `V27 -> "fedora-27"
   |`Fedora `V28 -> "fedora-28"
   |`Fedora `V29 -> "fedora-29"
+  |`Fedora `V30 -> "fedora-30"
   |`OracleLinux `V7 -> "oraclelinux-7"
   |`OracleLinux `Latest -> "oraclelinux"
   |`Alpine `V3_3 -> "alpine-3.3"
@@ -196,11 +204,13 @@ let tag_of_distro (d:t) = match d with
   |`Alpine `V3_7 -> "alpine-3.7"
   |`Alpine `V3_8 -> "alpine-3.8"
   |`Alpine `V3_9 -> "alpine-3.9"
+  |`Alpine `V3_10 -> "alpine-3.10"
   |`Alpine `Latest -> "alpine"
   |`OpenSUSE `V42_1 -> "opensuse-42.1"
   |`OpenSUSE `V42_2 -> "opensuse-42.2"
   |`OpenSUSE `V42_3 -> "opensuse-42.3"
   |`OpenSUSE `V15_0 -> "opensuse-15.0"
+  |`OpenSUSE `V15_1 -> "opensuse-15.1"
   |`OpenSUSE `Latest -> "opensuse"
 
 let distro_of_tag x : t option = match x with
@@ -220,6 +230,7 @@ let distro_of_tag x : t option = match x with
   |"debian-stable" -> Some (`Debian `Stable)
   |"debian-unstable" -> Some (`Debian `Unstable)
   |"debian-testing" -> Some (`Debian `Testing)
+  |"debian-10" -> Some (`Debian `V10)
   |"debian-9" -> Some (`Debian `V9)
   |"debian-8" -> Some (`Debian `V8)
   |"debian-7" -> Some (`Debian `V7)
@@ -234,6 +245,7 @@ let distro_of_tag x : t option = match x with
   |"fedora-27" -> Some (`Fedora `V27)
   |"fedora-28" -> Some (`Fedora `V28)
   |"fedora-29" -> Some (`Fedora `V29)
+  |"fedora-30" -> Some (`Fedora `V30)
   |"fedora" -> Some (`Fedora `Latest)
   |"oraclelinux-7" -> Some (`OracleLinux `V7)
   |"oraclelinux" -> Some (`OracleLinux `Latest)
@@ -244,11 +256,13 @@ let distro_of_tag x : t option = match x with
   |"alpine-3.7" -> Some (`Alpine `V3_7)
   |"alpine-3.8" -> Some (`Alpine `V3_8)
   |"alpine-3.9" -> Some (`Alpine `V3_9)
+  |"alpine-3.10" -> Some (`Alpine `V3_10)
   |"alpine" -> Some (`Alpine `Latest)
   |"opensuse-42.1" -> Some (`OpenSUSE `V42_1)
   |"opensuse-42.2" -> Some (`OpenSUSE `V42_2)
   |"opensuse-42.3" -> Some (`OpenSUSE `V42_3)
   |"opensuse-15.0" -> Some (`OpenSUSE `V15_0)
+  |"opensuse-15.1" -> Some (`OpenSUSE `V15_1)
   |"opensuse" -> Some (`OpenSUSE `Latest)
   |_ -> None
 
@@ -269,6 +283,7 @@ let rec human_readable_string_of_distro (d:t) =
   |`Debian `Stable -> "Debian Stable"
   |`Debian `Unstable -> "Debian Unstable"
   |`Debian `Testing -> "Debian Testing"
+  |`Debian `V10 -> "Debian 9 (Buster)"
   |`Debian `V9 -> "Debian 9 (Stretch)"
   |`Debian `V8 -> "Debian 8 (Jessie)"
   |`Debian `V7 -> "Debian 7 (Wheezy)"
@@ -283,6 +298,7 @@ let rec human_readable_string_of_distro (d:t) =
   |`Fedora `V27 -> "Fedora 27"
   |`Fedora `V28 -> "Fedora 28"
   |`Fedora `V29 -> "Fedora 29"
+  |`Fedora `V30 -> "Fedora 30"
   |`OracleLinux `V7 -> "OracleLinux 7"
   |`Alpine `V3_3 -> "Alpine 3.3"
   |`Alpine `V3_4 -> "Alpine 3.4"
@@ -291,10 +307,12 @@ let rec human_readable_string_of_distro (d:t) =
   |`Alpine `V3_7 -> "Alpine 3.7"
   |`Alpine `V3_8 -> "Alpine 3.8"
   |`Alpine `V3_9 -> "Alpine 3.9"
+  |`Alpine `V3_10 -> "Alpine 3.10"
   |`OpenSUSE `V42_1 -> "OpenSUSE 42.1"
   |`OpenSUSE `V42_2 -> "OpenSUSE 42.2"
   |`OpenSUSE `V42_3 -> "OpenSUSE 42.3"
   |`OpenSUSE `V15_0 -> "OpenSUSE 15.0 (Leap)"
+  |`OpenSUSE `V15_1 -> "OpenSUSE 15.1 (Leap)"
   |`Alpine `Latest | `Ubuntu `Latest | `Ubuntu `LTS | `CentOS `Latest | `Fedora `Latest
   |`OracleLinux `Latest | `OpenSUSE `Latest -> alias ()
 
@@ -343,6 +361,7 @@ let base_distro_tag d =
           | `V3_7 -> "3.7"
           | `V3_8 -> "3.8"
           | `V3_9 -> "3.9"
+          | `V3_10 -> "3.10"
           | `Latest -> assert false
         in
         "alpine", tag
@@ -352,6 +371,7 @@ let base_distro_tag d =
           | `V7 -> "7"
           | `V8 -> "8"
           | `V9 -> "9"
+          | `V10 -> "buster" (* TODO until final release *)
           | `Testing -> "testing"
           | `Unstable -> "unstable"
           | `Stable -> assert false
@@ -388,7 +408,8 @@ let base_distro_tag d =
           | `V26 -> "26"
           | `V27 -> "27"
           | `V28 -> "28"
-          | `V29 -> "28"
+          | `V29 -> "29"
+          | `V30 -> "30"
           | `Latest -> assert false
         in
         "fedora", tag
@@ -402,6 +423,7 @@ let base_distro_tag d =
           | `V42_2 -> "42.2"
           | `V42_3 -> "42.3"
           | `V15_0 -> "15.0"
+          | `V15_1 -> "15.1"
           | `Latest -> assert false
         in
         "opensuse/leap", tag